pyproject.toml
[tool.poetry]
name = "project"
version = "0.1.0"
description = ""
authors = []

[tool.poetry.dependencies]
python = "^3.7"
django = "^3.0.6"
django-extensions = "^2.2.9"
django-debug-toolbar = "^2.2"
werkzeug = "^1.0.1"
django-environ = "^0.4.5"
django-imagekit = "^4.0.2"
Pillow = "^7.2.0"
django-allauth = "^0.42.0"
djangorestframework = "^3.11.1"
django-registration-redux = "^2.8"
psycopg2 = "^2.8.5"
django-heroku = "^0.3.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
